Refactor to make generic search resource
